Move CompactionTaskStatusesBuilder to core
