+++ java/compaction/compaction-core/src/test/java/sleeper/compaction/core/task/CompactionTaskStatusesBuilderTest.java
@@ -21,6 +21,7 @@ import sleeper.core.record.process.RecordsProcessed;
 import sleeper.core.record.process.RecordsProcessedSummary;
 import sleeper.core.tracker.compaction.task.CompactionTaskFinishedStatus;
 import sleeper.core.tracker.compaction.task.CompactionTaskStatus;
+import sleeper.core.tracker.compaction.task.CompactionTaskStatusesBuilder;
 
 import java.time.Instant;
 import java.util.List;
