+++ java/core/src/main/java/sleeper/core/tracker/compaction/task/CompactionTaskStatusesBuilder.java
@@ -13,10 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.core.task;
-
-import sleeper.core.tracker.compaction.task.CompactionTaskFinishedStatus;
-import sleeper.core.tracker.compaction.task.CompactionTaskStatus;
+package sleeper.core.tracker.compaction.task;
 
 import java.time.Instant;
 import java.util.Comparator;
